+++ test/core/interface_name_shadow.v
@@ -1,3 +1,9 @@
+module A(input wire [3:0] i);
+    initial $display("A %b", i);
+endmodule
+module B(input wire [3:0] i);
+    initial $display("B %b", i);
+endmodule
 module top;
     generate
         if (1) begin : i
@@ -9,4 +15,7 @@ module top;
         end
     endgenerate
     initial $display("%b %b", i.x, blk.i);
+    A a(i.x);
+    B b(i.x);
+    assign i.x = 1;
 endmodule